Top action bar changes, reformatted how titles are set.
Action bar now has the basis for containing extended functionality.
Titles are now in res/strings.xml are are set in MainActivity.java
