+++ android/app/src/main/java/com/example/myfair/activities/CardCreationActivity.java
@@ -1,4 +1,4 @@
-package com.example.myfair;
+package com.example.myfair.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
+import com.example.myfair.R;
 import com.example.myfair.db.Card;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
