+++ android/app/src/main/java/com/example/myfair/activities/GenerateActivity.java
@@ -1,4 +1,4 @@
-package com.example.myfair;
+package com.example.myfair.activities;
 
 import android.content.Context;
 
@@ -19,6 +19,7 @@ import android.widget.Toast;
 import com.example.myfair.ModelsandHelpers.EncryptionHelper;
 import com.example.myfair.ModelsandHelpers.QRCodeHelper;
 import com.example.myfair.ModelsandHelpers.qrObject;
+import com.example.myfair.R;
 import com.google.gson.Gson;
 
 import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
