+++ android/app/src/main/java/com/example/myfair/activities/ScannedActivity.java
@@ -1,15 +1,14 @@
-package com.example.myfair;
+package com.example.myfair.activities;
 
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.example.myfair.ModelsandHelpers.EncryptionHelper;
 import com.example.myfair.ModelsandHelpers.qrObject;
+import com.example.myfair.R;
 import com.example.myfair.db.Card;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
