+++ android/app/src/main/res/layout/activity_card_creation.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".CardCreationActivity">
+    tools:context=".activities.CardCreationActivity">
 
     <Button
         android:id="@+id/btnDone"
