+++ android/app/src/main/res/layout/activity_generate.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".GenerateActivity">
+    tools:context=".activities.GenerateActivity">
