+++ android/app/src/main/res/layout/activity_profile_creation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".ProfileCreationActivity">
+    tools:context=".activities.ProfileCreationActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/lytTopLevel"
