+++ android/app/src/main/res/layout/activity_scan.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ScanActivity">
+    tools:context=".activities.ScanActivity">
 
 
     <me.dm7.barcodescanner.zxing.ZXingScannerView
