+++ android/app/src/main/res/layout/activity_scanned.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
     android:orientation="vertical"
-    tools:context=".ScannedActivity">
+    tools:context=".activities.ScannedActivity">
 
     <TextView
         android:layout_width="match_parent"
