+++ android/app/src/main/res/layout/activity_user_library.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".UserLibraryActivity">
+    tools:context=".activities.UserLibraryActivity">
 
     <TextView
         android:id="@+id/textView4"
