+++ android/app/src/main/res/layout/fragment_analytics.xml
@@ -5,7 +5,7 @@
     android:id="@+id/fragmentLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AnalyticsFragment">
+    tools:context=".fragments.AnalyticsFragment">
 
     <TextView
         android:id="@+id/textView3"
