+++ android/app/src/main/res/layout/fragment_create.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".CreateFragment">
+    tools:context=".fragments.CreateFragment">
 
     <TextView
         android:id="@+id/textView4"
