+++ android/app/src/main/res/layout/fragment_history.xml
@@ -7,7 +7,7 @@ xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".HistoryFragment">
+    tools:context=".fragments.HistoryFragment">
 
     <TextView
         android:id="@+id/textView5"
