+++ android/app/src/main/res/menu/actionmenu_collections.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <!-- "Mark Favorite", should appear as action button if possible -->
+    <item
+        android:id="@+id/action_menu_search"
+        android:icon="@drawable/ic_collections"
+        android:title="@string/action_menu_collections_search"
+        app:showAsAction="ifRoom"/>
+
+    <!-- Settings, should always be in the overflow -->
+    <item
+        android:id="@+id/action_menu_filter"
+        android:icon="@drawable/ic_history"
+        android:title="@string/action_menu_collections_filter"
+        app:showAsAction="ifRoom"/>
+
+</menu>
